Conflicts:
	cocos2d-ios.xcodeproj/project.pbxproj
	cocos2d-mac.xcodeproj/project.pbxproj
	cocos2d/CCLabelTTF.m
	cocos2d/CCNode.h
	cocos2d/CCNode.m
	cocos2d/CCSprite.m
	cocos2d/CCTexture2D.m
	cocoslive/CLScoreServerRequest.m
	tests/LabelTest.h
	tests/LabelTest.m

Former-commit-id: c3cc511e56abbec4c517178517d98f4cb6e30436

+++ cocos2d/Support/ccCArray.h
@@ -100,7 +100,7 @@ static inline void ccArrayEnsureExtraCapacity(ccArray *arr, NSUInteger extra)
 /** shrinks the array so the memory footprint corresponds with the number of items */
 static inline void ccArrayShrink(ccArray *arr)
 {
-    uint newSize;
+    NSUInteger newSize;
 	
 	//only resize when necessary
 	if (arr->max > arr->num && !(arr->num==0 && arr->max==1))
